refactor: stats, event 패키지 재분리 (#282)
