+++ src/main/java/com/soda/project/domain/ProjectService.java
@@ -14,10 +14,10 @@ import com.soda.project.domain.member.MemberProject;
 import com.soda.project.domain.enums.ProjectStatus;
 import com.soda.project.domain.error.ProjectErrorCode;
 import com.soda.project.domain.member.MemberProjectService;
-import com.soda.project.event.ProjectCreatedEvent;
+import com.soda.project.domain.event.ProjectCreatedEvent;
 import com.soda.project.interfaces.dto.*;
 import com.soda.project.domain.stage.StageService;
-import com.soda.project.stats.repository.ProjectDailyStatsRepository;
+import com.soda.project.infrastructure.ProjectDailyStatsRepository;
 import com.soda.project.infrastructure.ProjectRepository;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
