+++ src/main/java/com/soda/project/infrastructure/ProjectDailyStatsRepositoryImpl.java
@@ -1,4 +1,4 @@
-package com.soda.project.stats.repository;
+package com.soda.project.infrastructure;
 
 import com.querydsl.core.Tuple;
 import com.querydsl.core.types.dsl.Expressions;
